Refactor and rename the Schwarz classes.
+ schwarz_solver -> schwarz_base + restricted_schwarz
