+++ include/schwarz_base.hpp
@@ -32,8 +32,8 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<SCHWARZ LIB LICENSE>*************************/
 
 
-#ifndef schwarz_solver_hpp
-#define schwarz_solver_hpp
+#ifndef schwarz_base_hpp
+#define schwarz_base_hpp
 
 #include <cmath>
 #include <fstream>
@@ -41,9 +41,10 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <omp.h>
 
+#include <schwarz/config.hpp>
+
 #include <communicate.hpp>
 #include <initialization.hpp>
-#include <schwarz/config.hpp>
 #include <solve.hpp>
 
 
@@ -69,9 +70,9 @@ namespace SchwarzWrappers {
  */
 template <typename ValueType = gko::default_precision,
           typename IndexType = gko::int32>
-class SolverBase : public Initialize<ValueType, IndexType>,
-                   public Communicate<ValueType, IndexType>,
-                   public Solve<ValueType, IndexType> {
+class SchwarzBase : public Initialize<ValueType, IndexType>,
+                    public Communicate<ValueType, IndexType>,
+                    public Solve<ValueType, IndexType> {
 public:
     /**
      * The constructor that takes in the user settings and a metadata struct
@@ -80,7 +81,7 @@ public:
      * @param settings  The settings struct.
      * @param metadata  The metadata struct.
      */
-    SolverBase(Settings &settings, Metadata<ValueType, IndexType> &metadata);
+    SchwarzBase(Settings &settings, Metadata<ValueType, IndexType> &metadata);
 
 #if SCHW_HAVE_DEALII
     /**
@@ -197,74 +198,7 @@ protected:
 };
 
 
-/**
- * An implementation of the solver interface using the RAS solver.
- *
- * @tparam ValueType  The type of the floating point values.
- * @tparam IndexType  The type of the index type values.
- * @ingroup SchwarzWrappers
- */
-template <typename ValueType = gko::default_precision,
-          typename IndexType = gko::int32>
-class SolverRAS : public SolverBase<ValueType, IndexType> {
-public:
-    /**
-     * Standardized data struct to pipe additional data to the solver.
-     */
-    struct AdditionalData {};
-
-    /**
-     * The constructor that takes in the user settings and a metadata struct
-     * containing the solver metadata.
-     *
-     * @param settings  The settings struct.
-     * @param metadata  The metadata struct.
-     * @param data  The additional data struct.
-     */
-    SolverRAS(Settings &settings, Metadata<ValueType, IndexType> &metadata,
-              const AdditionalData &data = AdditionalData());
-
-    void setup_local_matrices(
-        Settings &settings, Metadata<ValueType, IndexType> &metadata,
-        std::vector<unsigned int> &partition_indices,
-        std::shared_ptr<gko::matrix::Csr<ValueType, IndexType>> &global_matrix,
-        std::shared_ptr<gko::matrix::Csr<ValueType, IndexType>> &local_matrix,
-        std::shared_ptr<gko::matrix::Csr<ValueType, IndexType>>
-            &interface_matrix,
-        std::shared_ptr<gko::matrix::Permutation<IndexType>> &local_perm,
-        std::shared_ptr<gko::matrix::Permutation<IndexType>> &local_inv_perm)
-        override;
-
-    void setup_comm_buffers() override;
-
-    void setup_windows(
-        const Settings &settings,
-        const Metadata<ValueType, IndexType> &metadata,
-        std::shared_ptr<gko::matrix::Dense<ValueType>> &main_buffer) override;
-
-    void exchange_boundary(const Settings &settings,
-                           const Metadata<ValueType, IndexType> &metadata,
-                           std::shared_ptr<gko::matrix::Dense<ValueType>>
-                               &solution_vector) override;
-
-    void update_boundary(
-        const Settings &settings,
-        const Metadata<ValueType, IndexType> &metadata,
-        std::shared_ptr<gko::matrix::Dense<ValueType>> &local_solution,
-        const std::shared_ptr<gko::matrix::Dense<ValueType>> &local_rhs,
-        const std::shared_ptr<gko::matrix::Dense<ValueType>> &solution_vector,
-        std::shared_ptr<gko::matrix::Dense<ValueType>> &global_old_solution,
-        const std::shared_ptr<gko::matrix::Csr<ValueType, IndexType>>
-            &interface_matrix) override;
-
-    /**
-     * Store a copy of the flags for this particular solver.
-     */
-    const AdditionalData additional_data;
-};
-
-
 }  // namespace SchwarzWrappers
 
 
-#endif  // schwarz_solver.hpp
+#endif  // schwarz_base.hpp
