+++ run_bench
@@ -1,0 +1,1 @@
+mpirun -n 4 ./benchmarking/bench_ras --explicit_laplacian --enable_global_check
